added get more dogs button functionality

+++ src/components/DogImages.js
@@ -3,18 +3,20 @@ import Dog from "./Dog"
 
 function DogImages() {
     const [images, setImages] = useState([])
+    const [send, setSend] = useState(false);
 
-    // grab 10 random images from dog ceo api
+    // grab 4 random images from dog ceo api
     useEffect(() => {
+        if (send === true) return;
         // delete any current data
         fetch("http://localhost:4000/images")
         .then(res => res.json())
         .then(data => {
             if (data.length === 0) return;
             data.forEach(dog => {
-            fetch(`http://localhost:4000/images/${dog.id}`, {
-                method: "DELETE"
-            })
+                fetch(`http://localhost:4000/images/${dog.id}`, {
+                    method: "DELETE"
+                })
             })
         })
 
@@ -22,38 +24,42 @@ function DogImages() {
         fetch("https://dog.ceo/api/breeds/image/random/4")
         .then(res => res.json())
         .then(async (data) => {
-  
-        // post new dog ceo images to json file
-        data.message.map(link => {
-            const dogNameArray = link.split("/")[4].split("-")
-            const reversed = [];
-            for (const i in dogNameArray) {
-                const capitalized = dogNameArray[i].toUpperCase().slice(0, 1) + dogNameArray[i].toLowerCase().slice(1)
-                reversed.unshift(capitalized)
-            }
-            const dogName = reversed.join(" ")
-            const desc = "Random picture of " + dogName
-  
-            const dogObj = {
-              image: link,
-              name: dogName,
-              description: desc
-            }
-            
-            fetch("http://localhost:4000/images", {
-            method: "POST",
-            headers: {
-              "Content-Type": "application/json",
-            },
-            body: JSON.stringify(dogObj)
-          })
-        })})
+            // post new dog ceo images to json file
+            data.message.map(link => {
+                const dogNameArray = link.split("/")[4].split("-")
+                const reversed = [];
+                for (const i in dogNameArray) {
+                    const capitalized = dogNameArray[i].toUpperCase().slice(0, 1) + dogNameArray[i].toLowerCase().slice(1)
+                    reversed.unshift(capitalized)
+                }
+                const dogName = reversed.join(" ")
+                const desc = "Random picture of " + dogName
+    
+                const dogObj = {
+                image: link,
+                name: dogName,
+                description: desc
+                }
+                
+                fetch("http://localhost:4000/images", {
+                method: "POST",
+                headers: {
+                "Content-Type": "application/json",
+                },
+                body: JSON.stringify(dogObj)
+                })
+            })
+        })
 
         // add dog objects to state
         fetch("http://localhost:4000/images")
         .then(res => res.json())
         .then(data => setImages(data))
-    }, [])
+
+        return () => {
+            setSend(false)
+        }
+    }, [send])
 
     // create a Dog component for each image
     const displayImages = images.map(image => {
@@ -64,7 +70,7 @@ function DogImages() {
 
     return (
         <div>
-            <button className="dog-button">Get More Dogs</button>
+            <button onClick={() => setSend(true)} className="dog-button">Get More Dogs</button>
             <div className="dog-container">
                 {displayImages}
             </div>
